abstract project creation
